+++ src/views/LostItems/index.js
@@ -1,2 +1,2 @@
-export { default as NoPermissionMessage } from './NoPermissionMessage';
+// eslint-disable-next-line import/prefer-default-export
 export { default as LostItemsListContainer } from './LostItemsListContainer';