+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-users
 
-## [11.0.0] IN PROGRESS
+## [11.0.0](https://github.com/folio-org/ui-users/tree/v11.0.0) (2024-10-31)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v10.1.2...v11.0.0)
 
 * UX consistency: Use Save & close button label stripes-component translation key. Refs UIU-3078.
 * Fix two "triangle down" icons in select element of "Copy existing fee/fine owner table entries" modal. Refs UIU-2929.
@@ -73,6 +74,10 @@
 * Create new user record from preregistration data when no matching folio user is found. Refs UIU-3223.
 * Update permission name after Review and cleanup Module Descriptor for ui-checkout. Refs UIU-3216.
 * Integrate action to merge preregistration patron record with existing user record. Refs UIU-3225.
+* Update translations for pre registrations value. Refs UIU-3241.
+* Display a message to a user without permission on the “Patron preregistration record results” page. Refs UIU-3242.
+* Display preregistration data appropriately. Refs UIU-3247.
+* Process more item tokens in due date print. Refs UIU-3239.
 * Review and cleanup Module Descriptors for ui-users. Refs UIU-3214.
 
 ## [10.1.2](https://github.com/folio-org/ui-users/tree/v10.1.2) (2024-09-05)